Some UI update (Panel scrolling pb)
Now the panel is of a fixes length. If we add too much nodes in the schedule or too much filtres, we can scroll them, without the high part.
However we need now to define a radio event for onWindows resize to udpate the UI, if not, tags we look weird.